+++ theme/src/data/theme-config.js
@@ -126,7 +126,8 @@ const defaultConfig = {
           title: 'Instagram'
         }
       ]
-    }
+    },
+    footer: []
   },
 
   // Widget configuration